Extract useful functions from processor to util
