+++ src/util.rs
@@ -1,6 +1,6 @@
 use regex::Regex;
 
-use crate::processor::{add_tone, modify_letter, remove_tone, LetterModification, ToneMark};
+use crate::{processor::{add_tone, modify_letter, remove_tone, LetterModification, ToneMark}, maps::{ACCUTE_MAP, GRAVE_MAP, HOOK_ABOVE_MAP, TILDE_MAP, DOT_MAP, HORN_MAP, BREVE_MAP, CIRCUMFLEX_MAP, DYET_MAP}};
 
 pub fn clean_char(ch: char) -> char {
     let accents = vec![
@@ -114,4 +114,59 @@ pub fn remove_tone_or_append(input: &mut String, append_char: &char) {
         // Append the trigger char if there's no tone to remove for input
         input.push(*append_char);
     }
+}
+
+const VOWELS: [char; 12] = ['a', 'ă', 'â', 'e', 'ê', 'i', 'o', 'ô', 'ơ', 'u', 'ư', 'y'];
+const MODIFIED_VOWELS: [char; 6] = ['ă', 'â', 'ê', 'ô', 'ơ', 'ư'];
+const MODIFIABLE_VOWELS: [char; 4] = ['a', 'e', 'o', 'u'];
+
+pub fn is_vowel(c: char) -> bool {
+    VOWELS.contains(&c)
+}
+
+pub fn is_modified_vowels(c: char) -> bool {
+    MODIFIED_VOWELS.contains(&c)
+}
+
+pub fn is_modifiable_vowels(c: char) -> bool {
+    MODIFIABLE_VOWELS.contains(&c)
+}
+
+pub fn extract_tone(input: &str) -> Option<ToneMark> {
+    for ch in input.chars() {
+        if ACCUTE_MAP.values().find(|c| **c == ch).is_some() {
+            return Some(ToneMark::Acute);
+        }
+        if GRAVE_MAP.values().find(|c| **c == ch).is_some() {
+            return Some(ToneMark::Grave);
+        }
+        if HOOK_ABOVE_MAP.values().find(|c| **c == ch).is_some() {
+            return Some(ToneMark::HookAbove);
+        }
+        if TILDE_MAP.values().find(|c| **c == ch).is_some() {
+            return Some(ToneMark::Tilde);
+        }
+        if DOT_MAP.values().find(|c| **c == ch).is_some() {
+            return Some(ToneMark::Underdot);
+        }
+    }
+    None
+}
+
+pub fn extract_letter_modification(input: &str) -> Option<LetterModification> {
+    for ch in input.chars() {
+        if HORN_MAP.values().find(|c| **c == ch).is_some() {
+            return Some(LetterModification::Horn);
+        }
+        if BREVE_MAP.values().find(|c| **c == ch).is_some() {
+            return Some(LetterModification::Breve);
+        }
+        if CIRCUMFLEX_MAP.values().find(|c| **c == ch).is_some() {
+            return Some(LetterModification::Circumflex);
+        }
+        if DYET_MAP.values().find(|c| **c == ch).is_some() {
+            return Some(LetterModification::Dyet);
+        }
+    }
+    None
 }
